test git

+++ src/main/java/org/example/exercice5spring/service/FurnitureService.java
@@ -3,6 +3,7 @@ package org.example.exercice5spring.service;
 import org.example.exercice5spring.dao.FurnitureRepository;
 import org.example.exercice5spring.entity.Furniture;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 
@@ -23,19 +24,21 @@ public class FurnitureService {
     }
 
     public Furniture getFurnitureById(Long id) {
-        return repository.findById(id).orElse(null);
+        return repository.findById(id).orElseThrow(() -> new RuntimeException("Furniture not found"));
     }
 
+    @Transactional
     public void deleteFurnitureById(Long id) {
         if (id == null) {
             throw new IllegalArgumentException("The given id must not be null");
+        }
+        if (!repository.existsById(id)) {
+            throw new RuntimeException("Furniture not found");
         }
         repository.deleteById(id);
     }
 
     public void updateFurniture(Furniture furniture) {
-        repository.save(furniture);
+        saveFurniture(furniture);
     }
-
-
 }
